requirements.txt
Flask==3.1.2
Flask-CORS==6.0.1
psycopg2-binary==2.9.11
python-dotenv==1.1.1
PyJWT==2.10.1
cryptography==46.0.3
requests==2.32.5
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.7.2
matplotlib==3.10.7
scipy==1.16.2
gunicorn==23.0.0